Your work was mentioned in the book
Hi, Yoshitani!

I'm very sorry, but I did not find any other way to let you know, that this plugin was reviewed in my "Mastering Redmine" book. Also you was mentioned there as the one, who updated the plugin to work with recent Redmine versions.

See this link:
http://projects.andriylesyuk.com/projects/mastering-redmine/wiki/Mentioned-projects-and-people#Mentioned-people-and-organizations

Certainly, you do not need to approve this file change! I used it only to send the message to you.

Thanks.
Andriy